Shifting some of the redirects/rewrites above App
